User still be able to use the app if they have previously logged in.
if we have had a user logged in and out on a mobile device. If they come
to login again but we cannot access parse.com or no data network in
mobile, they can still login if we have the user they are trying to in
local storage as mentioned above

Done this point please check.